eliminates redundant patient.service
